Add support for pulling ietf-yang-library instead of capabilities (#26)
* Capture device schema to a git repo.

* Capture device schema to a git repo.

* Minor updates to print statements.

* Add support for generating per-platform metadata.

This metadata describes the platform, OS version, OS flavor, and any
tested PIDs.  While here, I made capabilities file generation
per-platform type.

* Make sure the model name does not contain spaces.

* Add support for generating per-platform metadata. (#1)

* Add support for generating per-platform metadata.

This metadata describes the platform, OS version, OS flavor, and any
tested PIDs.  While here, I made capabilities file generation
per-platform type.

* Make sure the model name does not contain spaces.

* Add support for NX-OS.

* Add NX-OS support (#2)

* Add support for generating per-platform metadata.

This metadata describes the platform, OS version, OS flavor, and any
tested PIDs.  While here, I made capabilities file generation
per-platform type.

* Make sure the model name does not contain spaces.

* Add support for NX-OS.

* Add support for getting the repo directory and owner

The per-platform metadata needs to the repo directory name and owner in
order to be useful in the GitHub API.

* Make sure the capabilities-file.path is relative to the repo dir.

* Add some additional metadata parameters (#4)

* Add support for generating per-platform metadata.

This metadata describes the platform, OS version, OS flavor, and any
tested PIDs.  While here, I made capabilities file generation
per-platform type.

* Make sure the model name does not contain spaces.

* Add support for NX-OS.

* Add support for getting the repo directory and owner

The per-platform metadata needs to the repo directory name and owner in
order to be useful in the GitHub API.

* Make sure the capabilities-file.path is relative to the repo dir.

* Do not normalize NX-OS versions in the platform metadata.

* Do not normalize NX-OS versions in the platform metadata. (#5)

* Add support for generating per-platform metadata.

This metadata describes the platform, OS version, OS flavor, and any
tested PIDs.  While here, I made capabilities file generation
per-platform type.

* Make sure the model name does not contain spaces.

* Add support for NX-OS.

* Add support for getting the repo directory and owner

The per-platform metadata needs to the repo directory name and owner in
order to be useful in the GitHub API.

* Make sure the capabilities-file.path is relative to the repo dir.

* Do not normalize NX-OS versions in the platform metadata.

* Change repo to repository.

* Add support for an ncclient device-type.

The new --device-type argument allows one to specify a device name to
ncclient.  This is needed for NX-OS devices.  By default, no device-type
is used, so backwards compatibility is preserved.

* Replace None with an empty dict.

* Add an explicit 'type' field to the capabilities file.

In the future, this may be changed when ietf-yang-library support is
added.

* Rename capabilities-file to module-file-list.

This adds consistency to the current API.

* Rename models to product-ids to avoid confusion.

* Add support for pulling the ietf-yang-library.

If the ietf-yang-library is supported, it will be pulled from the device
and its data will be written to the capabilities file.  Else, the
fallback will be to grab the capabilities.

* Make sure we properly fallback to capabilities in all cases.